+++ recognition/45811935_Project/train.py
@@ -230,72 +230,6 @@ training_results_vqvae = pd.read_csv(VQVAE_WEIGHTS_PATH + 'training_vqvae.log', 
 plot_vqvae_train_val_results(training_results_vqvae)
 
 
-# Generate and plot some reconstructions from test set, along with their codes and inputs and SSIMs
-
-
-def generate_vqvae_images(images):
-    """
-    Plots and saves the input, codebook, and reconstructions made by the VQ-VAE, of the given 
-    images. Also calculates and saves the SSIM between each image and its reconstruction.
-    
-    Args:
-        images: images to pass through the VQ-VAE
-    """
-
-    # Reconstruction
-    reconstructed = vqvae.predict(images)
-    
-    # Code (Flattened)
-    encoder_outputs = vqvae.get_encoder().predict(images)
-    encoder_outputs_flattened = encoder_outputs.reshape(-1, encoder_outputs.shape[-1])
-    codebook_indices = vqvae.get_vq().get_codebook_indices(encoder_outputs_flattened)
-    
-    # Code (reshaped)
-    codebook_indices = codebook_indices.numpy().reshape(encoder_outputs.shape[:-1])
-    
-    # Unshift and reshape reconstructions and original images and print SSIM values
-    plt.figure()
-    for i in range(len(images)):
-        test_image = tf.reshape(images[i], (1, IMG_SHAPE, IMG_SHAPE, num_channels))\
-                         + PIXEL_SHIFT
-    
-        reconstructed_image = tf.reshape(reconstructed[i],
-                                         (1, IMG_SHAPE, IMG_SHAPE, num_channels)) + PIXEL_SHIFT
-    
-        codebook_image = codebook_indices[i]
-    
-        plt.subplot(1, 3, 1)
-        plt.imshow(tf.squeeze(test_image))
-        plt.title("Test Image")
-        plt.axis("off")
-    
-        plt.subplot(1, 3, 2)
-        plt.imshow(codebook_image)
-        plt.title("Codebook")
-        plt.axis("off")
-    
-        plt.subplot(1, 3, 3)
-        plt.imshow(tf.squeeze(reconstructed_image))
-        plt.title("Reconstruction")
-        plt.axis("off")
-    
-        plt.savefig(RESULTS_PATH + f'vq_vae_reconstructions_{i}.png')
-    
-        ssim = tf.math.reduce_sum(tf.image.ssim(test_image, reconstructed_image,
-                                                max_val=1.0)).numpy()
-    
-        main_stdout = sys.stdout
-        with open(RESULTS_PATH + 'main_results.txt', 'a') as f:
-            sys.stdout = f
-            print(f"SSIM between Test Image {i} and Reconstruction: ", ssim)
-            sys.stdout = main_stdout
-
-
-for sample_batch in test_data.take(1).as_numpy_iterator():
-    sample_batch = sample_batch[:NUM_IMAGES_TO_SHOW]
-    generate_vqvae_images(sample_batch)
-
-
 # Print overall mean test SSIM for VQ-VAE to file:
 with tf.device(device):
     test_mean_ssim_vqvae = mean_ssim(test_data, num_test_images, vqvae)
